#Kishor Naik : As per you comment i had added Decimal data type in User id Property..

+++ GithubBranchingStratergy/GithubBranchingStratergy/KishorAdmin/UserEntity.cs
@@ -8,7 +8,7 @@ namespace KishorAdmin
 {
     public class UserEntity
     {
-        public int UserId { get; set; }
+        public decimal UserId { get; set; }
 
         public String Name { get; set; }
     }
